requirements.txt
apiai==1.2.3
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
dialogflow==0.1.0
enum34==1.1.6
Flask==0.12.2
futures==3.2.0
google-api-core==0.1.3
google-auth==1.2.1
googleapis-common-protos==1.5.3
grpcio==1.8.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.13.3
pkg-resources==0.0.0
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
